Create separate file for browserSniffing utils
